Easily enable/disable tmux in bash
Also, this commit enables colored prompts in alacritty.

+++ .bashrc
@@ -37,7 +37,7 @@ fi
 
 # set a fancy prompt (non-color, unless we know we "want" color)
 case "$TERM" in
-    xterm-color|*-256color) color_prompt=yes;;
+    xterm-color|*-256color|alacritty) color_prompt=yes;;
 esac
 
 # uncomment for a colored prompt, if the terminal has the capability; turned
@@ -202,21 +202,32 @@ alias cs='cht.sh --shell'
 ###################################################################
 # Tmux startup customization.
 ###################################################################
-# Make a shorter alias for tmux.
-alias t='tmux'
-# Make a shorter alias to kill all the other sessions.
-alias tka='tmux kill-session -a'
-# Update tmux source file at startup.
-tmux source-file ~/.tmux.conf
-# Run tmux at startup and test that (1) tmux exists on the system, (2) it is in an interactive shell, and (3) tmux doesn't try to run within itself.
-if command -v tmux &> /dev/null && [ -n "$PS1" ] && [[ ! "$TERM" =~ screen ]] && [[ ! "$TERM" =~ tmux ]] && [ -z "$TMUX" ]; then
-    # Find the terminal name. Could be alacritty, konsole, gnome-terminal, etc.
-    current_terminal=$(ps -o 'cmd=' -p "$(ps -o 'ppid=' -p $$)")
-    # Match the current terminal path to any string ending with "alacritty".
-    # This is useful so that tmux only runs inside alacritty.
-    if [[ $current_terminal == *"alacritty" ]]; then
-        # Attach tmux if a session exists, otherwise run tmux.
-        tmux attach || exec tmux
+# Enable or disable tmux at startup.
+is_tmux_enabled=false
+# Whether or not to load tmux configurations.
+if $is_tmux_enabled ; then
+    # Make a shorter alias for tmux.
+    alias t='tmux'
+    # Make a shorter alias to kill all the other sessions.
+    alias tka='tmux kill-session -a'
+    # Update tmux source file at startup.
+    tmux source-file ~/.tmux.conf
+    # Run tmux at startup and test that (1) tmux exists on the system, (2) it
+    # is in an interactive shell, and (3) tmux doesn't try to run within
+    # itself.
+    if command -v tmux &> /dev/null \
+        && [ -n "$PS1" ] \
+        && [[ ! "$TERM" =~ screen ]] \
+        && [[ ! "$TERM" =~ tmux ]] \
+        && [ -z "$TMUX" ]; then
+        # Find the terminal name. Could be alacritty, konsole, etc.
+        current_terminal=$(ps -o 'cmd=' -p "$(ps -o 'ppid=' -p $$)")
+        # Match the current terminal path to any string ending with "alacritty".
+        # This is useful so that tmux only runs inside alacritty.
+        if [[ $current_terminal == *"alacritty" ]]; then
+            # Attach tmux if a session exists, otherwise run tmux.
+            tmux attach || exec tmux
+        fi
     fi
 fi
 ###################################################################
